Use code instead of a binding to create "articles"
The binding that creates the badge value is not working correctly when "articles" is created using another binding. The combo of the two of them seem to be breaking. It makes the binding value of the badge to be updated by increasing the value with the new length.

Instead of being set to 4 it increases 4.
This part of the binding "@rssController.articles.filter{!isRead}" is an always increasing array.
When "articles" is not created by a binding everything works as expected.